DeleteItem + DeleteList + confirms done

+++ src/components/DeleteItem/DeleteItem.js
@@ -3,22 +3,22 @@ import DeleteIcon from '@mui/icons-material/Delete';
 
 export function DeleteItem(props) {
 
-function handleClickDelete() {
-    console.log(props.title)
-    console.log(props.index)
-    console.log(props.listName)
+    function handleClickDeleteItem() {
+        const result = confirm(`Are you sure you want to delete ${props.title}?`);
+        if (result) {
 
-    const dataFromStorage = JSON.parse(localStorage.getItem("lists")) || {}
-    const listFromStorage = dataFromStorage[props.listName]
-    // filter below should splice be replaced
-    // listFromStorage.filter(el=>el.title !== props.title)
-    listFromStorage.splice(props.index, 1)
+            const dataFromStorage = JSON.parse(localStorage.getItem("lists")) || {}
+            const listFromStorage = dataFromStorage[props.listName]
+            // filter below should splice be replaced
+            // listFromStorage.filter(el=>el.title !== props.title)
+            listFromStorage.splice(props.index, 1)
 
-    localStorage.setItem('lists', JSON.stringify(dataFromStorage))
-    location.reload()
-}
+            localStorage.setItem('lists', JSON.stringify(dataFromStorage))
+            location.reload()
+        }
+    }
 
     return (
-        <DeleteIcon onClick={handleClickDelete}/>
+        <DeleteIcon onClick={handleClickDeleteItem}/>
     )
 }
